requirements.txt
grpcio
grpcio-tools
protobuf
torch
transformers
diffusers
accelerate
imageio
imageio-ffmpeg
gradio
huggingface_hub
spaces
git+https://github.com/openai/whisper.git
opencv-python
numpy
matplotlib